requirements.txt
moderngl
moderngl-window
numpy
pillow
matplotlib
streamlit>=1.28.0
pyyaml
scipy
imageio
